requirements.txt
# Runtime deps for AppleBBCH81 conversion/util scripts

pillow>=9.0.0

# Optional (for COCO API example usage)
# pycocotools>=2.0.7


